feat: use EnterpriseContractPolicy custom resource
This refactors the tests to create and use EnterpriseContractPolicy
custom resource instead of the ConfigMap. In verify-enterprise-contract
Tekton Task both ConfigMap and EnterpriseContractPolicy are supported,
whereas in v2 only EnterpriseContractPolicy is supported.

The tests now assert the state as expected today, i.e. the failing Tasks
are asserted to fail as we expect today. We no longer assert the
`OUTPUT` TaskRun result as that is brittle with corresponding changes to
the changes in hacbs-contract/ec-policy repository.

By default the policy sources are fetched from hacbs-contract/ec-policy
repository main branch. This can be overridden by creating a ConfigMap
e2e-tests/ec-config with keys `revision` and `repository`, each of those
can replace this default settings. This will come in useful if we want
to pin a certain revision we know the tests are good to run against,
e.g. if we enable pull request updates via
redhat-appstudio/infra-deployments from hacbs-contract/ec-policies; or
if we setup a separate repository with policies suited for the end to
end tests. This brings the groundwork for that, either of those options
could be a good solution to reduce the brittleness, but this is left for
the future.

Also refactors the tests to add more debug output when the tests fail.