refactor: bundle image reference for pull requests
Pull request bundle images are from
`quay.io/repository/redhat-appstudio/pull-request-builds` repository and
they have tags reflecting their flavor (base or HACBS). The replacement
that was looking for `/<flavor>` was not appropriate for this as the
flavor was in the tag (e.g. `:<flavor>`). This adjusts for both pull
request builds that change the bundles and those that do not.
It also seems that `HACBSCoreServiceTemplatesBundle` is unused, so
instead of adjusting that the logic to deduce the value and the field
has been removed.